+++ tests/base/model_train_dataloaders.py
@@ -20,7 +20,7 @@ from tests.base.dataloaders import CustomNotImplementedErrorDataloader
 class TrainDataloaderVariations(ABC):
 
     @abstractmethod
-    def dataloader(self, train: bool):
+    def dataloader(self, train: bool, *args, **kwargs):
         """placeholder"""
 
     def train_dataloader(self):
@@ -37,3 +37,12 @@ class TrainDataloaderVariations(ABC):
         dataloader.dataset.data = dataloader.dataset.data[:0]
         dataloader.dataset.targets = dataloader.dataset.targets[:0]
         return dataloader
+
+    def train_dataloader__multiple_mapping(self):
+        """Return a mapping loaders with different lengths"""
+        return {'a': self.dataloader(train=True, num_samples=100),
+                'b': self.dataloader(train=True, num_samples=50)}
+    
+    def train_dataloader__multiple_sequence(self):
+        return [self.dataloader(train=True, num_samples=100),
+                self.dataloader(train=True, num_samples=50)]